changed data type of Camera field 'fov' from int to float

+++ src/org/xjge/core/Camera.java
@@ -22,7 +22,7 @@ import org.joml.Vector3f;
  */
 public abstract class Camera {
     
-    protected int fov = 45;
+    protected float fov = 45.0f;
     
     boolean isOrtho;
